Move fetching in SearchList
